NFDIV-4060 Move duplicated code to helper
