+++ src/main/java/uk/gov/hmcts/divorce/caseworker/event/CaseworkerHwfApplicationAccepted.java
@@ -11,15 +11,9 @@ import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
 import uk.gov.hmcts.divorce.divorcecase.model.State;
 import uk.gov.hmcts.divorce.divorcecase.model.UserRole;
 
-import java.time.Clock;
-import java.time.LocalDateTime;
-
-import static uk.gov.hmcts.ccd.sdk.type.YesOrNo.NO;
-import static uk.gov.hmcts.ccd.sdk.type.YesOrNo.YES;
 import static uk.gov.hmcts.divorce.divorcecase.model.State.AwaitingDocuments;
 import static uk.gov.hmcts.divorce.divorcecase.model.State.AwaitingHWFDecision;
 import static uk.gov.hmcts.divorce.divorcecase.model.State.AwaitingPayment;
-import static uk.gov.hmcts.divorce.divorcecase.model.State.Submitted;
 import static uk.gov.hmcts.divorce.divorcecase.model.UserRole.CASE_WORKER;
 import static uk.gov.hmcts.divorce.divorcecase.model.UserRole.JUDGE;
 import static uk.gov.hmcts.divorce.divorcecase.model.UserRole.LEGAL_ADVISOR;
@@ -32,7 +26,7 @@ public class CaseworkerHwfApplicationAccepted implements CCDConfig<CaseData, Sta
     public static final String CASEWORKER_HWF_APPLICATION_ACCEPTED = "caseworker-hwf-application-accepted";
 
     @Autowired
-    private Clock clock;
+    private CaseworkerHwfApplicationAndPaymentHelper caseworkerHwfApplicationAndPaymentHelper;
 
     @Override
     public void configure(final ConfigBuilder<CaseData, State, UserRole> configBuilder) {
@@ -52,21 +46,8 @@ public class CaseworkerHwfApplicationAccepted implements CCDConfig<CaseData, Sta
 
         CaseData caseData = details.getData();
 
-        if (caseData.getApplicationType().isSole()
-            && NO.equals(caseData.getApplication().getApplicant1KnowsApplicant2Address())
-            && YES.equals(caseData.getApplication().getApplicant1WantsToHavePapersServedAnotherWay())) {
-            details.setState(AwaitingDocuments);
-        } else {
-            details.setState(Submitted);
-        }
-
-        if (null == caseData.getApplication().getDateSubmitted()) {
-            caseData.getApplication().setDateSubmitted(LocalDateTime.now(clock));
-        }
-
-        if (null == caseData.getDueDate()) {
-            caseData.setDueDate(caseData.getApplication().getDateOfSubmissionResponse());
-        }
+        details.setState(caseworkerHwfApplicationAndPaymentHelper.getState(caseData));
+        details.setData(caseworkerHwfApplicationAndPaymentHelper.setDateSubmittedAndDueDate(caseData));
 
         return AboutToStartOrSubmitResponse.<CaseData, State>builder()
             .data(details.getData())
